+++ components/Testimonials.js
@@ -1,15 +1,18 @@
+import Image from "next/image";
+import { PROJECT_NAME } from "@constants/index";
+
 import Link from "./Link";
 
 export default function Testimonials({ data }) {
   return (
-    <div className="bg-white dark:bg-primary-high pt-0 pb-24">
+    <div className="bg-white dark:bg-dark pb-24 pt-10">
       <div className="mx-auto max-w-7xl px-6 lg:px-8">
         <div className="mx-auto max-w-xl text-center">
-          <h2 className="text-lg font-semibold leading-8 tracking-tight text-secondary-medium dark:text-secondary-low">
+          <h2 className="font-semibold text-secondary-high dark:text-secondary-low text-3xl">
             Testimonials
           </h2>
           <p className="mt-2 text-3xl font-bold tracking-tight text-primary-high dark:text-primary-low">
-            What people are saying about LinkFree...
+            What people are saying about {PROJECT_NAME}...
           </p>
         </div>
         <div className="mx-auto mt-16 flow-root">
@@ -21,10 +24,12 @@ export default function Testimonials({ data }) {
                     <p>{`“${testimonial.text}”`}</p>
                   </blockquote>
                   <figcaption className="mt-6 flex items-center gap-x-4">
-                    <img
+                    <Image
                       className="h-10 w-10 rounded-full bg-primary-low"
                       src={testimonial.image}
                       alt={`Profile picture for ${testimonial.name}`}
+                      width={100}
+                      height={100}
                     />
                     <div>
                       <Link
@@ -33,7 +38,9 @@ export default function Testimonials({ data }) {
                       >
                         {testimonial.name} ({`@${testimonial.username}`})
                       </Link>
-                      <div className="text-primary-medium dark:text-primary-low-medium">{testimonial.bio}</div>
+                      <div className="text-primary-medium dark:text-primary-low-medium">
+                        {testimonial.bio}
+                      </div>
                     </div>
                   </figcaption>
                 </figure>