+++ components/user/UserTestimonials.js
@@ -1,17 +1,17 @@
-import { ReactMarkdown } from "react-markdown/lib/react-markdown";
-
 import Link from "@components/Link";
 import Alert from "@components/Alert";
 import FallbackImage from "@components/FallbackImage";
+import Markdown from "@components/Markdown";
 
-export default function UserTestimonials({ data }) {
+export default function UserTestimonials({ testimonials, BASE_URL }) {
   return (
     <>
-      {!data.testimonials && (
-        <Alert type="info" message="No testimonials found" />
+      {!testimonials?.length && (
+        <Alert type="info" message="No Testimonials found" />
       )}
-      {data.testimonials &&
-        data.testimonials.map((testimonial, key) => (
+
+      {testimonials &&
+        testimonials.map((testimonial, key) => (
           <div
             className="flex flex-col sm:flex-row sm:gap-8 gap-2 sm:items-center text-sm dark:text-primary-low dark:bg-primary-medium text-primary-medium-low dark:border-none border-2 my-4 px-5 p-6 rounded-xl shadow-xl"
             key={key}
@@ -22,6 +22,7 @@ export default function UserTestimonials({ data }) {
                   width={100} // just to pass nextjs error
                   height={100}
                   src={`https://github.com/${testimonial.username}.png`}
+                  fallback={testimonial.username}
                   alt={testimonial.username}
                   className="rounded-full bg-primary-low"
                 />
@@ -30,7 +31,10 @@ export default function UserTestimonials({ data }) {
                 <h3 className="font-medium dark:text-primary-medium-low text-primary-high">
                   {testimonial.title}
                 </h3>
-                <Link href={testimonial.url} target="_blank">
+                <Link
+                  href={`${BASE_URL}/${testimonial.username}`}
+                  target="_blank"
+                >
                   @{testimonial.username}
                 </Link>
               </div>
@@ -41,12 +45,15 @@ export default function UserTestimonials({ data }) {
                 <h3 className="font-medium dark:text-primary-low text-primary-high">
                   {testimonial.title}
                 </h3>
-                <Link href={testimonial.url} target="_blank">
+                <Link
+                  href={`${BASE_URL}/${testimonial.username}`}
+                  target="_blank"
+                >
                   @{testimonial.username}
                 </Link>
               </div>
               <div className="prose prose-sm max-w-none w-fit text-primary-medium dark:text-primary-medium-low ">
-                <ReactMarkdown>{testimonial.description}</ReactMarkdown>
+                <Markdown>{testimonial.description}</Markdown>
               </div>
             </div>
           </div>