+++ pages/_app.js
@@ -1,6 +1,6 @@
 import { SessionProvider } from "next-auth/react";
 import { ThemeProvider } from "next-themes";
-import { GoogleAnalytics } from "nextjs-google-analytics";
+import { useRouter } from "next/router";
 import { Analytics } from "@vercel/analytics/react";
 
 import "../styles/globals.css";
@@ -10,6 +10,7 @@ export default function MyApp({
   Component,
   pageProps: { session, ...pageProps },
 }) {
+  const router = useRouter();
   // Use the layout defined at the page level, if available
   const getLayout =
     Component.getLayout || ((page) => <MultiLayout>{page}</MultiLayout>);
@@ -19,10 +20,10 @@ export default function MyApp({
       <SessionProvider session={session}>
         {getLayout(
           <>
-            <GoogleAnalytics trackPageViews />
-            <Component {...pageProps} />
+            <Component {...pageProps} key={router.asPath} />
             <Analytics />
-          </>
+          </>,
+          pageProps.settings
         )}
       </SessionProvider>
     </ThemeProvider>