+++ pages/api/account/statistics.js
@@ -3,10 +3,7 @@ import { getServerSession } from "next-auth/next";
 
 import connectMongo from "@config/mongo";
 import logger from "@config/logger";
-import Profile from "@models/Profile";
-import ProfileStats from "@models/ProfileStats";
-import Link from "@models/Link";
-import LinkStats from "@models/LinkStats";
+import { Profile, ProfileStats, Link, LinkStats } from "@models/index";
 
 export default async function handler(req, res) {
   const session = await getServerSession(req, res, authOptions);
@@ -22,19 +19,47 @@ export default async function handler(req, res) {
       .json({ error: "Invalid request: GET request required" });
   }
 
-  const username = session.username;
+  const data = await getStats(session.username);
+
+  res.status(200).json(data);
+}
+
+export async function getStats(username, numberOfDays = 30) {
   await connectMongo();
+
+  // This calculates the start date by subtracting the specified number of days from the current date. 
+  // The query then retrieves data from that calculated start date up to the current date.
+  const startDate = new Date();
+  startDate.setDate(startDate.getDate() - numberOfDays);
 
   let profileData = {};
   try {
-    profileData = await Profile.findOne({ username });
+    const rankedProfiles = await Profile.aggregate([
+      {
+        $setWindowFields: {
+          sortBy: { views: -1 },
+          output: {
+            rank: {
+              $rank: {},
+            },
+          },
+        },
+      },
+      {
+        $match: {
+          username: username,
+        },
+      },
+    ]);
+
+    profileData = rankedProfiles[0];
   } catch (e) {
     logger.error(e, "failed to load profile");
   }
 
   let profileViews = [];
   try {
-    profileViews = await ProfileStats.find({ username }).sort({ date: "asc" });
+    profileViews = await ProfileStats.find({ username, date: { $gte: startDate }}).sort({ date: "asc" });
   } catch (e) {
     logger.error(e, "failed to load stats");
   }
@@ -67,8 +92,8 @@ export default async function handler(req, res) {
   let totalClicks = 0;
   const linkDailyStats = linkClicks.map((item) => {
     totalClicks += item.clicks;
-
     return {
+      _id: item._id,
       url: item.url,
       clicks: item.clicks,
       daily: dailyClicks
@@ -85,6 +110,7 @@ export default async function handler(req, res) {
       total: profileData.views,
       monthly: totalViews,
       daily: dailyStats,
+      rank: profileData.rank,
     },
     links: {
       clicks: totalClicks,
@@ -92,5 +118,5 @@ export default async function handler(req, res) {
     },
   };
 
-  res.status(200).json(data);
+  return JSON.parse(JSON.stringify(data));
 }