+++ tests/events.spec.js
@@ -1,22 +1,32 @@
 // @ts-check
 import { test, expect } from "@playwright/test";
-const AxeBuilder = require("@axe-core/playwright").default;
+import AxeBuilder from "@axe-core/playwright";
 
 test("Click on events profile in navbar navigates to events page", async ({
   page,
 }) => {
   await page.goto("/");
-  await page.getByRole("link", { name: 'Events', exact: true }).click();
-  await expect(page).toHaveURL("/events");
+  await page
+    .getByRole("navigation")
+    .getByRole("link", { name: "Events" })
+    .click();
+  await page.waitForLoadState("networkidle");
+  await expect(page).toHaveURL(/\/events/);
 });
 
-test.fixme("Events listed", async ({ page }) => {
+test("Events has title", async ({ page }) => {
   await page.goto("/events");
-  await expect(page.locator("li")).toBeGreaterThan(1);
+  await expect(page.locator("h1")).toHaveText("Community Events");
+});
+
+test("Events listed", async ({ page }) => {
+  await page.goto("/events");
+  const elements = await page.locator("li").count();
+  await expect(elements).toBeGreaterThan(1);
 });
 
 test.describe("accessibility tests (light)", () => {
-  test.use({ colorScheme: 'light' });
+  test.use({ colorScheme: "light" });
 
   test("should pass axe wcag accessibility tests", async ({ page }) => {
     await page.goto("/events");
@@ -28,8 +38,8 @@ test.describe("accessibility tests (light)", () => {
 });
 
 test.describe("accessibility tests (dark)", () => {
-  test.use({ colorScheme: 'dark' });
-  
+  test.use({ colorScheme: "dark" });
+
   test("should pass axe wcag accessibility tests (dark)", async ({ page }) => {
     await page.goto("/events");
     const accessibilityScanResults = await new AxeBuilder({ page })