+++ tests/search.spec.js
@@ -1,6 +1,8 @@
 // @ts-check
 import { test, expect } from "@playwright/test";
-const AxeBuilder = require("@axe-core/playwright").default;
+import AxeBuilder from "@axe-core/playwright";
+
+const defaultUsers = 9;
 
 test("Search has title", async ({ page }) => {
   await page.goto("/search");
@@ -9,7 +11,11 @@ test("Search has title", async ({ page }) => {
 
 test("Navigate to the Search page", async ({ page }) => {
   await page.goto("/");
-  await page.locator("a:visible", { hasText: "Search" }).click();
+  await page
+    .getByRole("navigation")
+    .getByRole("link", { name: "Search" })
+    .click();
+  await page.waitForLoadState("networkidle");
   await expect(page.locator("h1")).toHaveText("Search");
 });
 
@@ -18,16 +24,16 @@ test("Search works correctly", async ({ page }) => {
   await page.goto("/search");
 
   // 2. show no users are listed
-  await expect(page.locator("li")).toHaveCount(0);
+  await expect(page.locator("main li")).toHaveCount(defaultUsers);
 
   // 3. type in search and check that user with the name exist and check a name doesn't exist
   const input = page.locator("[name='keyword']");
   await input.type("_test-profile-user-1");
 
-  await expect(page.locator("li")).toHaveCount(1);
+  await expect(page.locator("main li")).toHaveCount(1);
 });
 
-test("Search page has no results when no search term used", async ({
+test("Search page has random results when no search term used", async ({
   page,
 }) => {
   await page.goto("/search");
@@ -35,18 +41,18 @@ test("Search page has no results when no search term used", async ({
   const input = page.locator("[name='keyword']");
   await input.type("");
 
-  await expect(page.locator("li")).toHaveCount(0);
+  await expect(page.locator("main li")).toHaveCount(defaultUsers);
 });
 
-test("Search page shows no results after typing 2 characters", async ({
+test("Search page shows random results after typing 1 characters", async ({
   page,
 }) => {
   await page.goto("/search");
 
   const input = page.locator("[name='keyword']");
-  await input.type("ed");
+  await input.type("e");
 
-  await expect(page.locator("li")).toHaveCount(0);
+  await expect(page.locator("main li")).toHaveCount(defaultUsers);
 });
 
 test("Search page shows results after typing 3 characters", async ({
@@ -57,18 +63,30 @@ test("Search page shows results after typing 3 characters", async ({
   const input = page.locator("[name='keyword']");
   await input.type("aka");
 
-  await expect(page.locator("li")).toContainText(["aka"]);
+  await expect(page.locator("main li")).toContainText(["aka"]);
 });
 
-test.fixme("After search click profile", async ({ page }) => {
-  // 1. perform search
-  // 2. click on searched profile
-  // 3. check profile is displayed
+test("After search click profile", async ({ page }) => {
+  // 1. Perform search
+  await page.goto("/search");
+  const input = page.locator("[name='keyword']");
+  await input.type("eddiejaoude");
+
+  // 2. Click on the searched profile
+  const profileLinkSelector = 'a[href="/eddiejaoude"]';
+  const profileLink = page.locator(profileLinkSelector);
+  await profileLink.click();
+  await page.waitForLoadState("networkidle");
+
+  // 3. Check if the profile is displayed
+  const profileHeader = page.locator("h1");
+  const profileHeaderText = await profileHeader.innerText();
+  await expect(profileHeaderText).toContain("Eddie Jaoude"); 
 });
 
 test.fixme(
   "find the profile after providing concise name",
-  async ({ page }) => {
+  async () => {
     // 1. click on search profile
     // 2. type the whole name
     // 3. Display the profile if the name is correct
@@ -76,7 +94,7 @@ test.fixme(
 );
 
 test.describe("accessibility tests (light)", () => {
-  test.use({ colorScheme: 'light' });
+  test.use({ colorScheme: "light" });
 
   test("should pass axe wcag accessibility tests (light)", async ({ page }) => {
     await page.goto("/search");
@@ -88,7 +106,7 @@ test.describe("accessibility tests (light)", () => {
 });
 
 test.describe("accessibility tests (dark)", () => {
-  test.use({ colorScheme: 'dark' });
+  test.use({ colorScheme: "dark" });
 
   test("should pass axe wcag accessibility tests (dark)", async ({ page }) => {
     await page.goto("/search");